requirements.txt
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-readers-file
chromadb
fastmcp>=3.0.0
llama-index-tools-mcp
sentence-transformers
